First version of MatchingEngine and refactoring of double to BigDecimal and more

+++ src/main/java/org/team27/stocksim/model/market/Instrument.java
@@ -1,13 +1,15 @@
 package org.team27.stocksim.model.market;
 
+import java.math.BigDecimal;
+
 public abstract class Instrument {
     protected final String symbol;
     protected final String name;
-    protected final double tickSize;
+    protected final BigDecimal tickSize;
     protected final int lotSize;
 
 
-    public Instrument(String symbol, String name, double tickSize, int lotSize) {
+    public Instrument(String symbol, String name, BigDecimal tickSize, int lotSize) {
         this.symbol = symbol;
         this.name = name;
         this.tickSize = tickSize;
@@ -15,9 +17,9 @@ public abstract class Instrument {
     }
     public String getSymbol() {return symbol;}
     public String getName() {return name;}
-    public double getTickSize() {return tickSize;}
+    public BigDecimal getTickSize() {return tickSize;}
     public int getLotSize() {return lotSize;}
 
-    public abstract double getCurrentPrice();
+    public abstract BigDecimal getCurrentPrice();
 
 }
