+++ src/main/java/org/team27/stocksim/model/market/StockFactory.java
@@ -1,8 +1,10 @@
 package org.team27.stocksim.model.market;
 
+import java.math.BigDecimal;
+
 public class StockFactory implements InstrumentFactory{
     @Override
-    public Instrument createInstrument(String symbol, String name, Double tickSize, int lotSize) {
+    public Instrument createInstrument(String symbol, String name, BigDecimal tickSize, int lotSize) {
         return new Stock(symbol, name, tickSize, lotSize);
     }
 }
